Extract code from postcss printer (#14480)
